Cambios

+++ libreria/proyectfmake.h
@@ -1,6 +1,5 @@
-
-float distancia(float x1,float x2);
-float fuerza(float r,float x,float m);
-float velocidad(float vxp,float fxp,float fx,float h);
-float posicion(float xp,float h,float vxp,float fxp);
-float distanciat(float dx,float dy,float dz);
+double distancia(double x1,double x2);
+double fuerza(double r,double x,double m);
+double velocidad(double vxp,double fxp,double fx,double h);
+double posicion(double xp,double h,double vxp,double fxp);
+double distanciat(double dx,double dy,double dz);
